requirements.txt
absl-py==1.1.0
ale-py==0.7.5
alembic==1.8.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
autopage==0.5.1
beautifulsoup4==4.10.0
black==22.6.0
cachetools==5.2.0
certifi==2022.5.18.1
click==8.1.3
cliff==3.10.1
cloudpickle==2.1.0
cmaes==0.8.2
cmd2==2.4.1
colorlog==6.6.0
debugpy==1.6.0
docker-pycreds==0.4.0
empyrical==0.5.5
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.7.1
finta==1.3
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.47.0
gym==0.21.0
gym-anytrading==1.3.1
gym-notices==0.0.7
h5py==3.7.0
huggingface-hub==0.8.1
imageio==2.19.3
importlib-resources==5.8.0
ipykernel==6.10.0
ipython==8.2.0
ipywidgets==7.7.0
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.1
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
lxml==4.9.1
Mako==1.2.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multitasking==0.0.11
munkres==1.1.4
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.8.3
notebook==6.4.10
oauthlib==3.2.0
omegaconf==2.0.6
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
optuna==2.10.1
pandas==1.4.1
pandas-datareader==0.10.0
pathspec==0.9.0
pathtools==0.1.2
pbr==5.9.0
platformdirs==2.5.2
plotly==5.6.0
prettytable==3.3.0
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.0
pyperclip==1.8.2
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.18.1
pytorch-metric-learning==1.5.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.6.2
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sentencepiece==0.1.96
sentry-sdk==1.6.0
setproctitle==1.2.3
shortuuid==1.0.9
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.1
SQLAlchemy==1.4.39
stable-baselines3==1.5.0
stevedore==3.5.0
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tokenizers==0.12.1
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.20.1
vidaug==1.5
wandb==0.12.21
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
yfinance==0.1.73
